# specs/001-report-download-endpoint/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Log Analyzer Download API
  version: 1.0.0
  description: API para listar logs y descargar reportes por nombre, formato y archivos.
servers:
  - url: http://localhost:8080
paths:
  /datasets:
    get:
      summary: Listar archivos de logs disponibles
      responses:
        "200":
          description: Lista de logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetList"
  /reports/download:
    post:
      summary: Generar y descargar un reporte
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportDownloadRequest"
      responses:
        "200":
          description: Archivo generado para descarga
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          description: Solicitud invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Archivos no encontrados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    DatasetList:
      type: object
      properties:
        status:
          type: string
          example: success
        datasets:
          type: array
          items:
            type: string
        count:
          type: integer
    ReportDownloadRequest:
      type: object
      properties:
        report_name:
          type: string
          example: reporte-semana-1
        format:
          type: string
          enum: [excel, txt, csv, doc]
        files:
          type: array
          items:
            type: string
          example: ["generated_logs.txt", "app_errors.txt"]
      required:
        - report_name
        - format
        - files
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        error:
          type: string